Convertertendo para DaoImplements.

+++ src/main/java/br/senai/sc/dao/Dao.java
@@ -12,7 +12,7 @@ public interface Dao<T> {
 	
 	public void deleteAll(List<T> t);
 	
-	public T findAllById(Integer id);
+	public T findAllById(Long id);
 	
 	public List<T> findAll();
